feat/refactor: refactor navigation links and moved into a new Menu component, as well as implemented menu design for tablet-mobile devices
