+++ src/components/Navigation.tsx
@@ -1,17 +1,22 @@
-"use client";
-
 import "@/styles/nav.css";
-import NavItem from "./NavItem";
 import ModalWithForm from "./ModalWithForm";
+import Menu from "./Menu";
 import { robotoSlab } from "@/vendor/fonts";
-import { ReactNode, useState } from "react";
-import Login from "./Login";
+import { ReactNode, useEffect, useState } from "react";
+import { usePathname } from "next/navigation";
+import Image from "next/image";
 
 export default function Navigation() {
+  const pathname = usePathname();
+  const [menu, setMenu] = useState(false as boolean);
   const [popup, setPopup] = useState<{
     title: string;
     children: ReactNode;
   } | null>(null);
+
+  useEffect(() => {
+    document.addEventListener("keydown", () => setMenu(false));
+  }, []);
 
   const openPopup = (popup: { title: string; children: ReactNode }) => {
     setPopup(popup);
@@ -26,33 +31,41 @@ export default function Navigation() {
     if (e.key === "Escape") closePopup();
   };
 
-  const loginPopup = {
-    title: "Iniciar sesión",
-    children: <Login />,
-  };
-
   return (
     <>
-      <div className="navigation flex w-full h-[80px] px-[104px] absolute top-0 border-b-1 border-current/20">
+      <div
+        className={`navigation flex w-full h-[80px] px-[104px] absolute top-0 border-b-1 border-current/20 ${
+          menu && "border-white/20"
+        }`}
+      >
         <span
-          className={`navigation__title ${robotoSlab.className} font-bold`}
+          className={`navigation__title ${robotoSlab.className} font-bold ${
+            menu && "text-white"
+          }`}
           onClick={() => (window.location.href = "/")}
         >
           News Explorer
         </span>
         <nav className="navigation__menu hidden md:flex font-medium h-full">
-          <ul className="navigation__list flex h-full">
-            <NavItem title="Inicio" link="/" />
-            <NavItem title="Artículos guardados" link="/saved-news" />
-          </ul>
-          <button
-            className="navigation__button w-[176px] p-[12px] rounded-full hover:bg-current transition-all duration-300 font-medium"
-            onClick={() => openPopup(loginPopup)}
-          >
-            <span className="navigation__button-text">Iniciar sesión</span>
-          </button>
+          <Menu openPopup={openPopup} />
         </nav>
+
+        <Image
+          src={`${menu ? "/close.svg" : "/menu.svg"}`}
+          alt="Menu/Close icon"
+          width={24}
+          height={24}
+          className={`md:hidden ${pathname != "/" && !menu && "invert"}`}
+          onClick={() => setMenu(!menu)}
+        />
       </div>
+
+      {menu && (
+        <nav className="navigation__menu-mobile flex flex-col w-full md:hidden font-medium">
+          <Menu openPopup={openPopup} />
+        </nav>
+      )}
+
       {popup && (
         <ModalWithForm
           title={popup.title}
