+++ src/styles/modal.css
@@ -9,7 +9,7 @@
   position: fixed;
   top: 0;
   left: 0;
-  z-index: 2;
+  z-index: 3;
   opacity: 1;
   transition: opacity 0.5s ease;
   @starting-style {
@@ -66,9 +66,11 @@
   }
 
   .modal__container {
+    padding: 25px 20px;
     width: 100%;
     height: 520px;
     border-radius: 16px 16px 0 0;
+    /* transition: height 1s; */
   }
 
   .modal__close {
@@ -76,11 +78,5 @@
     top: -32px;
     width: 24px;
     height: 24px;
-  }
-}
-
-@media screen and (max-width: 420px) {
-  .modal__container {
-    padding: 16px;
   }
 }
